+++ src/StandardIteratorInterface.php
@@ -95,6 +95,24 @@ interface StandardIteratorInterface extends IteratorAggregate, Countable
     public function map(callable $callback);
 
     /**
+     * @param $amount
+     * @return static
+     */
+    public function take($amount);
+
+    /**
+     * @param callable $callback
+     * @return static
+     */
+    public function tap(callable $callback);
+
+    /**
+     * @param callable $callback
+     * @return static
+     */
+    public function sort(callable $callback);
+
+    /**
      * @return bool
      */
     public function isEmpty();