+++ tests/StandardIteratorTest.php
@@ -184,4 +184,27 @@ class StandardIteratorTest extends TestCase
         $this->assertEquals($dates, DateTimeIterator::cast($other));
         $this->assertEquals($dates, DateTimeIterator::cast([$date]));
     }
+
+    public function testTap()
+    {
+        $tapped = null;
+
+        $this->iterator->tap(function (StandardIterator $tap) use (&$tapped) {
+            $this->assertNotSame($this->iterator, $tap);
+            $tapped = $tap->take(2);
+        });
+
+        $this->assertSame(['a', 'b'], $tapped->all());
+    }
+
+    public function testSort()
+    {
+        $actual = $this->iterator->sort(function ($a, $b) {
+            return strcmp($b, $a);
+        });
+
+        $this->assertNotSame($this->iterator, $actual);
+        $this->assertSame(['c', 'b', 'a'], $actual->all());
+        $this->assertSame(['a', 'b', 'c'], $this->iterator->all(), 'original is unaffected');
+    }
 }